backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# AI APIs
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.18.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.17.0
python-docx==1.1.0

# WebSocket & Real-time
websockets==12.0
redis==5.0.1
python-socketio==5.11.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# Utils
python-dotenv==1.0.1
pillow==10.2.0
python-magic==0.4.27
pytz==2024.1

# Logging & Monitoring
loguru==0.7.2